feat(analytics): move Top Dogs panel from /home to /analytics' Network tab
Task 1 of the Session 2 IA migration plan. Moves TopDogsPanel (and its
sub-components MetricCard, TierRow, plus TIER_CONFIG/METRIC_CONFIG/
TIERS_ORDER) verbatim from home/HomeOverview/index.jsx into
analytics/NetworkTab/index.jsx, wires fetch_global_performance_rankings()
into NetworkTab's existing data-fetching useEffect (via Promise.all
alongside fetch_node_geolocation), and renders it behind
PanelGate panelKey="topDogs" below the continent breakdown.

Removes the moved code and now-dead imports (FiCpu, FiHardDrive,
FiDownload, FiUpload, FaTrophy) from HomeOverview, confirmed unused
elsewhere in the file by grep. Also moves the associated .td-*/
.hov-panel--top-dogs CSS to NetworkTab/index.scss (not covered by
existing panel styling there).

Home.jsx's own separate fetch_global_performance_rankings() call and
globalRankings state are left intact per the brief - Task 2 removes
that dead fetch once all four panels have moved.
