+++ client/src/analytics/NetworkTab/index.scss
@@ -214,3 +214,130 @@
     color: #5eb8ff;
   }
 }
+
+// ── Top Dogs Panel ─────────────────────────────────────────────────────────────
+// Moved verbatim from home/HomeOverview/index.scss (Session 2 IA migration) —
+// these classes are only used by TopDogsPanel, which now lives in this file.
+
+.hov-panel--top-dogs { border-left: 2px solid #d97706; }
+.hov-panel--top-dogs .hov-header { background: linear-gradient(90deg, rgba(217, 119, 6, 0.04) 0%, transparent 60%); }
+
+.td-body { display: flex; flex-direction: column; gap: 10px; }
+
+.td-tier-row {
+  display: flex;
+  align-items: stretch;
+  gap: 10px;
+  padding: 8px 10px;
+  border-radius: 7px;
+  border-left: 3px solid transparent;
+  background: var(--surface-inset);
+}
+
+.td-tier-label-col {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 4px;
+  min-width: 52px;
+  flex-shrink: 0;
+}
+
+.td-tier-dot {
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+}
+
+.td-tier-name {
+  font-size: 0.65rem;
+  font-weight: 700;
+  letter-spacing: 0.04em;
+  text-transform: uppercase;
+}
+
+.td-metric-cards {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 8px;
+  flex: 1;
+  min-width: 0;
+
+  @media (max-width: 700px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+.td-metric-card {
+  display: flex;
+  flex-direction: column;
+  gap: 3px;
+  padding: 6px 8px;
+  border-radius: var(--radius-sm);
+  background: var(--surface-primary);
+  border: 1px solid var(--border-secondary);
+  min-width: 0;
+  overflow: hidden;
+  transition: border-color var(--transition-fast);
+
+  &:hover {
+    border-color: var(--border-hover);
+  }
+
+  &--empty {
+    opacity: 0.4;
+  }
+}
+
+.td-metric-header {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  color: var(--text-tertiary);
+}
+
+.td-metric-icon { flex-shrink: 0; }
+
+.td-metric-label {
+  font-size: 0.62rem;
+  font-weight: 700;
+  letter-spacing: 0.06em;
+  text-transform: uppercase;
+}
+
+.td-metric-ip {
+  font-size: 0.72rem;
+  font-variant-numeric: tabular-nums;
+  font-family: 'SFMono-Regular', 'Consolas', monospace;
+  color: var(--text-primary);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.td-metric-footer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 4px;
+}
+
+.td-metric-score {
+  font-size: 0.78rem;
+  font-weight: 700;
+  font-variant-numeric: tabular-nums;
+  color: #d97706;
+
+  @include rule-mode-dark() {
+    color: #fbbf24;
+  }
+}
+
+.td-header-icon {
+  color: #d97706;
+
+  @include rule-mode-dark() {
+    color: #fbbf24;
+  }
+}
